+++ compiler/src/iree/compiler/Codegen/Common/GPU/BUILD.bazel
@@ -63,7 +63,7 @@ iree_compiler_cc_library(
         "GPUApplyPaddingLevel.cpp",
         "GPUApplyTilingLevel.cpp",
         "GPUBubbleResourceCasts.cpp",
-        "GPUCanonicalizeIDs.cpp",
+        "GPUConvertAllIDsToThreadIDs.cpp",
         "GPUCheckResourceUsage.cpp",
         "GPUCombineLayoutTransformation.cpp",
         "GPUCombineValueBarriers.cpp",
