+++ compiler/src/iree/compiler/Codegen/LLVMGPU/Passes.cpp
@@ -956,8 +956,8 @@ addLowerAndOptimizeAddressComputationPasses(FunctionLikeNest &funcPassManager) {
       // full complexity.
       .addPass(createVectorTransferLoweringPass)
       .addPass(createIREECodegenFoldMemRefAliasOpsPass)
-      // Simplify gpu.*_id/dim ops based on annotated static info.
-      .addPass(createGPUCanonicalizeIDsPass)
+      // Rewrite all gpu.*_id ops in terms of thread ids.
+      .addPass(createGPUConvertAllIDsToThreadIDsPass)
       // Propagate constants close to loads/stores to improve the ability for
       // swizzling to CSE.
       .addPass(createPropagateConstantOffsetsPass)
